+++ CHANGELOG.md
@@ -22,6 +22,7 @@ k.vec2();
 - prevented default behavior for 'tab' and function keys
 - removed `loadAseprite()` and back to `loadSprite(name, src, { aseSpriteSheet: path, })`
 - supports newer version of pedit
+- added a simple plugin system, see the `multiboom` example
 
 # v0.3.0 "King Dedede...Bug!"
 - (**BREAK**) removed `pause()` and `paused()` in favor to `kaboom.debug.paused`